Wire up unit testing frameworks
There are so many, I don't really know which to choose from or why. I'm
just following some blogs for now. I've worked with a few of them
already and mocha and chai have come up along the way as well as jasmine
and karma. I'll start with these for now and possibly switch over later
if it makes sense.