Refactor types for reuseability.
Most of our types are following the same pattern of being a tree of
something -- it's just the something that changes. I've refactored the
types to all extend from a generic TreeLike and updated most of the
utility functions to work with any tree type. This should help us reuse
some of them for splitting the tree.