Implement function for splitting tree indexes
One of the main incentives for creating this as a library separate from
the core app was so that we'd have one place to centralize the logic for
both reading and writing the metadata files. The tree index is the least
trivial of the metadata files -- particularly since we want to be able
to split it up into smaller chunks that can be added incrementally -- so
this library was mainly built as a single place for building the logic
that serializes, deserializes, splits and recombines all of our metadata
representations along with some automated unit tests so that we can be
confident that what we've separated can be reconstituted back into the
original object that it came from.

This commit finally includes a rough working implementation of that
split and recombine code with some basic tests.