More WIP
The API is a total mess -- I should probably pull all of the ball tree
implementation into a completely separate set of utility functions. I've
implemented halfspace intersections. If we go anywhere with this, it
might also be time to drop the binary addressable indexed tree approach
for ball trees -- I'm not sure it really gives us anything.